Move financial verification to its own page
